requirements.txt
APScheduler==3.6.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
DataProperty==0.43.1
Deprecated==1.2.6
dominate==2.4.0
Flask==1.1.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mbstrdecoder==0.8.1
msgfy==0.0.7
pathvalidate==0.29.0
PyGithub==1.44
PyJWT==1.7.1
pymongo==3.9.0
pytablewriter==0.46.1
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
six==1.12.0
tabledata==0.9.1
typepy==0.6.0
tzlocal==2.0.0
urllib3==1.25.6
Werkzeug==0.16.0
wrapt==1.11.2
